+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # PASSWALL CHANGELOG
 
-## Version: [-.-.-] (2020-04-18)
+## Version: [1.0.8] (2020-04-18)
 ### Add
 - Search,Limit,Offset,Sort,Order query parameters to FindAll()
+- Backup to ./store/passwall.bak file
+- Restore from./store/passwall.bak file
   
 ### Changed
 - Order BT to updated_at DESC
 - Trim http://, https:// and www from URL's
+- Create util and helper folder.
+- Refactor
 
 ## Version: [1.0.7] (2020-04-17)
 ### Added